fix issue #304
