+++ package.json
@@ -33,15 +33,15 @@
   },
   "size-limit": [
     {
-      "limit": "1.675 KB",
+      "limit": "1.7 KB",
       "path": "lib/index.js"
     },
     {
-      "limit": "1.75 KB",
+      "limit": "1.8 KB",
       "path": "dist/react-swipeable.js"
     },
     {
-      "limit": "1.75 KB",
+      "limit": "1.8 KB",
       "path": "es/index.js"
     }
   ],
@@ -88,8 +88,8 @@
     "@types/jest": "^27.4.1",
     "@types/react": "^18.0.1",
     "@types/react-dom": "^18.0.0",
-    "@typescript-eslint/eslint-plugin": "^4.14.0",
-    "@typescript-eslint/parser": "^4.14.0",
+    "@typescript-eslint/eslint-plugin": "^5.20.0",
+    "@typescript-eslint/parser": "^5.20.0",
     "coveralls": "^3.0.3",
     "eslint": "^7.18.0",
     "eslint-config-prettier": "^6.11.0",
@@ -104,7 +104,7 @@
     "rollup": "^2.60.0",
     "size-limit": "^7.0.8",
     "ts-jest": "^27.1.4",
-    "typescript": "^4.1.3"
+    "typescript": "^4.6.3"
   },
   "dependencies": {},
   "peerDependencies": {
